Create machine-learning-and-managed-metadata
Metadata adoption is tough.  With the introduction of Machine Learning into the SharePoint ecosystem, Administrators and System Architects are asking if it can be leveraged to reduce our dependency on Users to assign metadata.  This article shares the results of an exploration into the technical capabilities of Machine Learning as it relates to SharePoint Managed Metadata.